Oops! more code to commit (to fix the previous bug...)
